+++ src/test/scala/astraea/spark/rasterframes/TestEnvironment.scala
@@ -32,7 +32,7 @@ trait TestEnvironment extends FunSpec with GeoTrellisTestEnvironment
 
   override implicit def sc: SparkContext = { _sc.setLogLevel("ERROR"); _sc }
 
-  lazy val sqlContext = {
+  lazy val sqlContext: SQLContext = {
     val ctx = SQLContext.getOrCreate(_sc).withRasterFrames
     ctx
   }